fix: Cairo perf output to a common root directory
I think the plan is to then have a pixelforge output
in the similar location.

+++ perf/test_line.mojo
@@ -16,7 +16,7 @@ comptime logger = Logger()
 def horizontal(
     context: ContextPtr, width: Int, height: Int, loops: Int
 ) -> PerfTime:
-    var h = height / 2 + 0.5
+    var h = Float64(height) / 2.0 + 0.5
     # move_to( context, 0, h)
     # line_to( context, width ,h)
